refactor: pkgフォルダをinternalフォルダに移動
