+++ internal/pkg/enum/connection_type.go
